added Blackjack, FizzBuzz, Hanoi, PrimeSieve, Sieve from Work

+++ src/Fibonacci.java
@@ -2,7 +2,7 @@ import java.util.Date;
 
 public class Fibonacci {
 
-    static long fibonacci(long n) {
+    static int fibonacci(int n) {
 
         if (n > 1)
             return fibonacci(n - 1) + fibonacci(n - 2);
@@ -13,7 +13,7 @@ public class Fibonacci {
     public static void main(String argv[]) {
 
         Date start = new Date();
-        long n = fibonacci(47);
+        int n = fibonacci(47);
         Date end = new Date();
 
         System.out.println(n);
